Create scoped_lock_lockingStraegies.cpp
explicit scoped_lock( MutexTypes&... m ); (1)	(since C++17)
scoped_lock( std::adopt_lock_t, MutexTypes&... m ); (2)	(since C++17)
scoped_lock( const scoped_lock& ) = delete;  (3)	(since C++17)

Acquires ownership of the given mutexes m.
1) If sizeof...(MutexTypes) == 0, does nothing. Otherwise, if sizeof...(MutexTypes) == 1, effectively calls m.lock(). Otherwise, effectively calls std::lock(m...).
2) Acquires ownership of the mutexes m... without attempting to lock any of them. The behavior is undefined unless the current thread holds a non-shared lock (i.e., a lock acquired by lock, try_lock, try_lock_for, or try_lock_until) on each object in m....
3) Copy constructor is deleted.
The behavior is undefined if m is destroyed before the scoped_lock object is.

scoped_lock is the combination of std::lock() function and lock_guard class.
It is the wrapper to the mutexes. It can simultaneously acquire the ownership of the multiple mutexes.
when the scoped_lock object is created, it try to acquire the ownership of the mutexes provided and lock them
and once the object goes out of scope, the scoped_lock object is destructed and the mutexes are released.

the behavior is undefined if the mutex objects are deleted before the scoped_object.